requirements.txt
flask
nltk
scikit-learn
pdfminer.six
